Twilight color fix: Binary build has 64-bit longs and 32-bit ints, GL c3i() takes an int[3] but was getting a long[3], messing up RGB. Changed Twilight colors to int.

+++ cedit/cedit.c
@@ -23,9 +23,9 @@ float ob = -1.0;
 int cc = -1;
 float cr, cg, cb;
 int m1, m2, m3;
-int xorg, yorg;
+long xorg, yorg;
 int x, y;
-int xsize, ysize;
+long xsize, ysize;
 float wx, wy;
 int menu;
